100% norminetted and working

+++ src/args.c
@@ -12,6 +12,7 @@
 
 #include "../includes/args.h"
 #include "../includes/minimap.h"
+#include "../includes/player.h"
 
 t_args	*init_args(mlx_image_t *img, mlx_t *mlx, t_map *map)
 {
@@ -26,7 +27,7 @@ t_args	*init_args(mlx_image_t *img, mlx_t *mlx, t_map *map)
 	args->player = init_player(map);
 	args->minimap = init_minimap(mlx, img, map);
 	args->step_angle = FOV / img->width;
-	args->scale = (args->img->width * 2) / CASTED_RAYS;
+	args->scale = (float)args->img->width * 2 / CASTED_RAYS;
 	args->texture[NO] = add_galloc(mlx_load_png(map->texture[NO]));
 	args->texture[EA] = add_galloc(mlx_load_png(map->texture[EA]));
 	args->texture[SO] = add_galloc(mlx_load_png(map->texture[SO]));
